A 2D action platform game for Android.
Portcullis Knight is 2D side-scrolling action platformer made for Android using SDL, meant to harken back to the 8 & 16 bit eras. This was started as a concept for a weekend game jam that became a fun project for me to practice SDL in Android and my wife to work on sprite art.